remove duplicate RingMock
